# Conflicts:
#	common/src/main/java/org/figuramc/figura/lua/api/FileAPI.java
#	common/src/main/java/org/figuramc/figura/lua/api/net/FiguraSocket.java
#	common/src/main/java/org/figuramc/figura/lua/api/net/HttpRequestsAPI.java
#	common/src/main/java/org/figuramc/figura/lua/api/net/NetworkingAPI.java
#	common/src/main/java/org/figuramc/figura/lua/api/net/SocketAPI.java

+++ common/src/main/java/org/figuramc/figura/lua/api/FileAPI.java
@@ -16,6 +16,7 @@ import org.luaj.vm2.LuaTable;
 import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -28,6 +29,7 @@ public class FileAPI {
     private final Avatar parent;
     private static final Path rootFolderPath = FiguraMod.getFiguraDirectory().resolve("data").toAbsolutePath()
             .normalize();
+    private static final String WRITE_NOT_ALLOWED = "You are only allowed to write in the data folder! Anything else is read only!";
 
     public FileAPI(Avatar parent) {
         this.parent = parent;
@@ -36,6 +38,7 @@ public class FileAPI {
     private Path securityCheck(String path) {
         if (!parent.isHost) throw new LuaError("You can't use FileAPI outside of host environment");
         Path p = relativizePath(path);
+        if (Files.isSymbolicLink(p)) throw new LuaError(String.format("Symbolic links are not allowed in FileAPI %s!", path));
         if (!isPathAllowed(p)) throw new LuaError(String.format("Path %s is not allowed in FileAPI", path));
         return p;
     }
@@ -60,7 +63,7 @@ public class FileAPI {
     }
 
     public boolean isPathAllowed(Path path) {
-        return path.toAbsolutePath().startsWith(rootFolderPath);
+        return !Files.isSymbolicLink(path) && path.toAbsolutePath().startsWith(rootFolderPath);
     }
 
     @LuaWhitelist
@@ -151,6 +154,9 @@ public class FileAPI {
     public FiguraOutputStream openWriteStream(@LuaNotNil String path) {
         try {
             Path p = securityCheck(path);
+            if (!p.startsWith(rootFolderPath)) {
+                throw new LuaError(WRITE_NOT_ALLOWED);
+            }
             File f = p.toFile();
             FileOutputStream fos = new FileOutputStream(f);
             return new FiguraOutputStream(fos);
@@ -261,6 +267,9 @@ public class FileAPI {
     )
     public boolean mkdir(@LuaNotNil String path) {
         Path p = securityCheck(path);
+        if (!p.startsWith(rootFolderPath)) {
+            throw new LuaError(WRITE_NOT_ALLOWED);
+        }
         File f = p.toFile();
         return f.mkdir();
     }
@@ -276,6 +285,9 @@ public class FileAPI {
     )
     public boolean mkdirs(@LuaNotNil String path) {
         Path p = securityCheck(path);
+        if (!p.startsWith(rootFolderPath)) {
+            throw new LuaError(WRITE_NOT_ALLOWED);
+        }
         File f = p.toFile();
         return f.mkdirs();
     }
@@ -291,6 +303,9 @@ public class FileAPI {
     )
     public boolean delete(@LuaNotNil String path) {
         Path p = securityCheck(path);
+        if (!p.startsWith(rootFolderPath)) {
+            throw new LuaError(WRITE_NOT_ALLOWED);
+        }
         File f = p.toFile();
         return f.delete();
     }
